tink-signature: move test to integration tests
